+++ ts/apps/polaris-cli/src/app/polaris-cli.impl.ts
@@ -5,14 +5,15 @@ import { hideBin } from 'yargs/helpers';
 import { PolarisCli } from './polaris-cli';
 import { DefaultTaskExecutor, TaskExecutor } from './tasks';
 import {
-    configCommand,
     createApplySloMappingCommand,
     createBuildCommand,
+    createConfigCommand,
     createDeployCommand,
     createDockerBuildCommand,
     createGenerateCommand,
     createGenerateCrdCommand,
     createInitCommand,
+    createMigrateCommand,
     createSerializeSloMappingCommand,
 } from './yargs-commands';
 
@@ -44,7 +45,8 @@ export class PolarisCliImpl implements PolarisCli {
             .command(createDockerBuildCommand(this))
             .command(createDeployCommand(this))
             .command(createGenerateCrdCommand(this))
-            .command(configCommand(this))
+            .command(createConfigCommand(this))
+            .command(createMigrateCommand(this))
             .help()
             .recommendCommands()
             .parseSync();