+++ ts/apps/polaris-cli/src/app/yargs-commands/index.ts
@@ -6,5 +6,6 @@ export * from './docker-build-command';
 export * from './generate-command';
 export * from './generate-crds-command';
 export * from './init-command';
+export * from './migrate-command';
 export * from './serialize-slo-mapping-command';
 export * from './apply-slo-mapping-command';