+++ ts/libs/polaris-nx/src/migrations/use-executor-for-docker-build/use-executor-for-docker-build.ts
@@ -1,0 +1,19 @@
+import { Tree, getProjects, updateProjectConfiguration } from '@nx/devkit';
+
+export default function updateDockerBuildCommand(tree: Tree) {
+    const projects = getProjects(tree);
+
+    for (const [name, project] of projects) {
+      if (
+        project.targets?.['docker-build']?.executor === 'nx:run-commands'
+      ) {
+        project.targets['docker-build'] = {
+            executor: '@polaris-sloc/polaris-nx:docker-build',
+            options: {
+                projectName: name,
+            },
+        };
+        updateProjectConfiguration(tree, name, project);
+      }
+    }
+}